Move is_dataset and is_soft_link to file_utils
